Change hw.physmem and hw.usermem to unsigned long like they used to be in the original hardwired sysctl implementation.
The buf size calculator still overflows an integer on machines with large
KVA (eg: ia64) where the number of pages does not fit into an int.  Use
'long' there.

Change Maxmem and physmem and related variables to 'long', mostly for
completeness.  Machines are not likely to overflow 'int' pages in the
near term, but then again, 640K ought to be enough for anybody.  This
comes for free on 32 bit machines, so why not?


git-svn-id: svn+ssh://svn.freebsd.org/base/head@102600 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/kern/subr_param.c
@@ -127,7 +127,7 @@ init_param1(void)
  * Boot time overrides that are scaled against main memory
  */
 void
-init_param2(int physpages)
+init_param2(long physpages)
 {
 
 	/* Base parameters */
